update classifcation script

+++ pot-mi/scripts/get_wb_classification.R
@@ -9,53 +9,55 @@ wbids = read.csv("data/ea_wbids.csv")
 
 #### FUNCTION TO GET SF FEATURES OF WATERBODY
 get_wb_classification = function(string, #### STRING = NAME OF CLASSFICATION AREA
-                     column)
+                                 column)
   #### COLUMN  = CLASSIFICATION TYPE E.G. OC | MC | RBD
 {
   #### LOGICAL OPERATOR FOR RIVER MINE
   if (column == "OC") {
-    wb = wbids %>% filter(OC == string)
+    wb = wbids %>% subset(OC == string)
   } # OPERATIONAL CATCHMENT
   if (column == "MC") {
-    wb = wbids %>% filter(MC == string)
+    wb = wbids %>% subset(MC == string)
   } # MANAGMENT CATCHMENT
   if (column == "RBD") {
-    wb = wbids %>% filter(RBD == string)
+    wb = wbids %>% subset(RBD == string)
   } # RIVER BASIN DISTRICT
+  if (column != "OC" & column != "MC" & column != "RBD") {
+    message("Woops, looks like you declared an invalid column type. Please try E.G. OC | MC | RBD")
+  } else{
+    #### LOOP THROUGH GEOJSON DOWNLOAD
+    suppressWarnings(for (i in 1:nrow(wb)) {
+      ##### EA CATCHMNET API CALL
+      url = "https://environment.data.gov.uk/catchment-planning/WaterBody/"
+      notation = wb$WBID[1]
+      download_url = paste0(url, notation, "/classifications.csv")
 
-  #### LOOP THROUGH GEOJSON DOWNLOAD
-  suppressWarnings(for (i in 1:nrow(wb)) {
-    ##### EA CATCHMNET API CALL
-    url = "https://environment.data.gov.uk/catchment-planning/WaterBody/"
-    notation = wb$WBID[i]
-    download_url = paste0(url, notation, "/classifications.csv")
+      #### SET OUTPUT PATH
+      river_wbid = wb$WBID[1]
+      path = "data/river_sf/"
+      river_output = paste0(path, river_wbid, ".csv")
 
-    #### SET OUTPUT PATH
-    river_wbid = wb$WBID[i]
-    path = "data/river_sf/"
-    river_output = paste0(path, river_wbid, ".csv")
+      #### DOWNLOOAD FILE, AT LEAST TRY TO
+      tryCatch(
+        expr = {
+          download.file(download_url, river_output)
+        },
+        error = function(e) {
+          message("Unable to download River Please check column and string are correct")
+        }
+      )
 
-    #### DOWNLOOAD FILE, AT LEAST TRY TO
-    tryCatch(
-      expr = {
-        download.file(download_url, river_output)
-      },
-      error = function(e) {
-        message("Unable to download River Please check column and string are correct")
-      }
-    )
+      river_sf = read.csv(river_output)
+      names(river_sf)[names(river_sf) == "Water.Body.ID"] = "WBID"
 
-    river_sf = read.csv(river_output) %>%  filter(Classification.Item == "Ecological") %>% filter(Year == 2019)
+      wb = merge(river_sf, wb)
 
-    wb$status[i] = river_sf$Status[1]
-    wb$year[i] = river_sf$Year[1]
-
-    #### REMOPVE DOWNLOADED FILE
-    file.remove(river_output)
-  })
-  return(wb)
+      #### REMOPVE DOWNLOADED FILE
+      file.remove(river_output)
+    })
+    return(wb)
+  }
 }
 
 # test function
-wb_class = get_wb_classification(string = "Thames",column = "RBD")
-
+wb_class = get_wb_classification(string = "River Till", column = "OC")
